+++ src/changelog.md
@@ -5,6 +5,9 @@ All notable changes to this documentation will be listed in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Calendar Versioning](https://calver.org).
 
+## Latest
+* Fix Community Container setup and start script
+
 ## 22.7.0 – 2022-07-25
 * Update docs for supporting 22.4 release